CLOUD RUN 2/requirements.txt
pandas
prophet
google-cloud-bigquery
google-cloud-storage
numpy==1.21.0
db-dtypes
